Fixed various warnings. Now resetting all delete object pointers to null to force error on use after delete.

+++ src/mapgui/maplayer.h
@@ -491,7 +491,7 @@ public:
     return airportMsaDetails;
   }
 
-  int getAirportMsaSymbolScale() const
+  float getAirportMsaSymbolScale() const
   {
     return airportMsaSymbolScale;
   }
